Corrected type in isSqrt.c

+++ Maths/Divisors/isSqrt/isSqrt.c
@@ -3,7 +3,7 @@
 
 //Gives 1 if n is a perfect square.
 int isSqrt(long long int n){
-    int i = sqrtN(n);
+    long int i = sqrtN(n);
     if(n==i*i)
         return 1;
     else
